+++ changelog.md
@@ -3,6 +3,7 @@
 ## Upcoming
 
 - added: post size now applies post text as well
+- added: seen on scroll delay
 - fixed: status bar disappearing
 - fixed: post card username pressable area
 - fixed: filters